Move `dat_to_csv` to `utils`
